+++ core/core-domain/src/main/java/com/barlow/core/domain/notificationsetting/NotificationSettingRepository.java
@@ -1,4 +1,4 @@
-package com.barlow.core.domain.legislationaccount.notificationsetting;
+package com.barlow.core.domain.notificationsetting;
 
 import java.util.List;
